Reorganize structure of node commands
