+++ src/node/satchecker/histogram.ts
